CR-120 Review changes
Marker interface introduced for event payload objects.
Route spelling error fixed.
Interface for EventPublisher removed as unnecessary.
