+++ src/main/java/uk/gov/ons/ctp/common/event/model/EventPayload.java
@@ -1,0 +1,4 @@
+package uk.gov.ons.ctp.common.event.model;
+
+/** Marker interface for event message detail sent as pay load */
+public interface EventPayload {}
